clothes_app/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
openai==0.27.8
requests==2.31.0
Werkzeug==2.3.4
Pillow==10.0.0
alembic==1.11.1